+++ freud/plot.py
@@ -1,3 +1,6 @@
+# Copyright (c) 2010-2019 The Regents of the University of Michigan
+# This file is from the freud project, released under the BSD 3-Clause License.
+
 import io
 import numpy as np
 try:
